editor stuff
saving and loading are in, game saves here:
C:\Users\[PC NAME]\AppData\Local\SmallPlatformer
PROBLEM: !!! Saving and loading complely breaks if we add new objects,
there is a way around this but it's horrible and painful. ( though i
think i'll do it anyways)
room that loads levels made in game in a playable way. (it's at the very
end, mash p to get there)